Refactor DataRows seeder into multiple files for each model  - Add ability to set icon in admin dimmer widgets  - Rebuild user_android_app table
