+++ database/seeds/DataRowsSeeders/MenusDataRowsTableSeeder.php
@@ -1,0 +1,72 @@
+<?php
+
+use Illuminate\Database\Seeder;
+use TCG\Voyager\Models\DataRow;
+use TCG\Voyager\Models\DataType;
+
+class MenusDataRowsTableSeeder extends BaseDataRowsTableSeeder
+{
+    /**
+     * Auto generated seed file.
+     */
+    public function run()
+    {
+        $dataType = DataType::where('slug', 'menus')->firstOrFail();
+
+        $dataRow = $this->dataRow($dataType, 'id');
+        $dataRow->fill([
+            'type' => 'number',
+            'display_name' => __('voyager::seeders.data_rows.id'),
+            'required' => 1,
+            'browse' => 0,
+            'read' => 0,
+            'edit' => 0,
+            'add' => 0,
+            'delete' => 0,
+            'details' => '',
+            'order' => 1,
+        ])->save();
+
+        $dataRow = $this->dataRow($dataType, 'name');
+        $dataRow->fill([
+            'type' => 'text',
+            'display_name' => __('voyager::seeders.data_rows.name'),
+            'required' => 1,
+            'browse' => 1,
+            'read' => 1,
+            'edit' => 1,
+            'add' => 1,
+            'delete' => 1,
+            'details' => '',
+            'order' => 2,
+        ])->save();
+
+        $dataRow = $this->dataRow($dataType, 'created_at');
+        $dataRow->fill([
+            'type' => 'timestamp',
+            'display_name' => __('voyager::seeders.data_rows.created_at'),
+            'required' => 0,
+            'browse' => 0,
+            'read' => 0,
+            'edit' => 0,
+            'add' => 0,
+            'delete' => 0,
+            'details' => '',
+            'order' => 3,
+        ])->save();
+
+        $dataRow = $this->dataRow($dataType, 'updated_at');
+        $dataRow->fill([
+            'type' => 'timestamp',
+            'display_name' => __('voyager::seeders.data_rows.updated_at'),
+            'required' => 0,
+            'browse' => 0,
+            'read' => 0,
+            'edit' => 0,
+            'add' => 0,
+            'delete' => 0,
+            'details' => '',
+            'order' => 4,
+        ])->save();
+    }
+}
